+++ WebDerm/src/Math/maths.h
@@ -27,7 +27,7 @@ namespace MyMath {
 	double euclideanDist(deque<double> &vec1, deque<double> &vec2);
 	double euclideanDist(double *vec1, double *vec2);
 	double euclideanDist(int *vec1, int *vec2);
-	double eucDist(Point pt1, Point pt2);
+	float eucDist(Point pt1, Point pt2);
 	double findDegreeDistance(double deg1, double deg2);
 
 	/* average.cpp */
